requirements.txt
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
Authlib==1.6.5
beartype==0.22.5
billiard==4.2.2
blinker==1.9.0
boto3==1.29.7
botocore==1.32.7
cachetools==6.2.1
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clipboard==0.0.4
colorama==0.4.6
connectors==4.1
contourpy==1.3.0
crispy-bootstrap5==2025.6
crispy-tailwind==0.5.0
cryptography==46.0.0
cycler==0.12.1
cyclopts==4.2.1
dacite==1.9.2
decouple==0.0.7
diskcache==5.6.3
distro==1.9.0
Django==4.2.7
django-ckeditor-5==0.2.12
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==3.2.3
django-rest-framework==0.1.0
django-storages==1.14.2
djangorestframework==3.15.1
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.22.2
dotenv==0.9.9
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastmcp==2.13.0.2
filelock==3.19.1
fonttools==4.54.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.38.0
greenlet==3.2.4
h11==0.16.0
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
ImageHash==4.3.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.6.0
kiwisolver==1.4.7
kombu==5.6.0
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mcp==1.20.0
mdurl==0.1.2
more-itertools==10.8.0
multimethod==1.12
narwhals==2.5.0
networkx==3.5
numba==0.61.0
numpy==2.1.2
openai==1.108.1
openapi-pydantic==0.5.1
openpyxl==3.1.5
oracledb==3.3.0
packaging==24.1
pandas==2.2.3
pathable==0.4.4
pathvalidate==3.3.1
patsy==1.0.1
phik==0.12.5
Pillow==10.0.1
platformdirs==4.4.0
plotly==6.3.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psycopg2-binary==2.9.10
puremagic==1.30
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.1.4
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2024.2
PyWavelets==1.9.0
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.8.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.17.3
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
sqlparse==0.5.0
sse-starlette==3.0.3
starlette==0.49.2
statsmodels==0.14.5
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.0.7
uvicorn==0.38.0
vine==5.1.0
visions==0.8.1
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
wordcloud==1.9.4
xlrd==2.0.2
ydata-profiling==4.16.0
